Fix e Feat: Correção do peso dos itens e da mochila de float para int. Adição da função que resolve o problema por programação dnâmica

+++ Problema-da-Mochila/item.c
@@ -11,7 +11,7 @@ float get_valor(ITEM* item)
     return (item->valorItem);
 }
 
-float get_peso(ITEM* item)
+int get_peso(ITEM* item)
 {
     return (item->pesoItem);
 }
