+++ Problema-da-Mochila/item.h
@@ -5,6 +5,6 @@
 typedef struct item_ ITEM;
 
 float get_valor(ITEM* item);
-float get_peso(ITEM* item);
+int get_peso(ITEM* item);
 bool item_apagar(ITEM** item);
 ITEM *item_criar(float peso, float valor);
